% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_values.R
\name{missing_values}
\alias{missing_values}
\title{Imputation of Missing Values}
\usage{
missing_values(inputdata, column.cutoff = NULL, group.cutoff = NULL,
  complete.matrix = FALSE, seed = 100, saveoutput = FALSE,
  outputname = "missing.values.rep")
}
\arguments{
\item{inputdata}{A data frame in the correct format.}

\item{column.cutoff}{A value between zero and one. If the proportion of
missing values is equal to or more than the column.cutoff in all groups, that
whole column will be deleted.}

\item{group.cutoff}{A value between zero and one. If the proportion of
missing values in a group is equal to or more than the group.cutoff, those
missing values will be replaced by a random number between zero and the
minimum of the entire matrix.}

\item{complete.matrix}{A logical indicating whether a complete matrix is
required. If \code{TRUE}, the remaining missing values (preferably only a
very few) will be replaced by the average of the abundances in the rest of
the group.}

\item{seed}{An integer to set the state of random number generation.}

\item{saveoutput}{A logical indicating whether the output should be saved.
If \code{TRUE}, the results will be saved as a csv file..}

\item{outputname}{The name of the output file.}
}
\value{
returns an object with class \code{metabdata}.
}
\description{
\code{missing_values()} imputes missing values in metabolmics data matrices
}
\note{
This function was copied from the now deprecated \code{metabolomics}
package. Authors listed below.
}
\seealso{
\href{https://github.com/cran/metabolomics}{metabolomics}
}
\author{
Alysha M De Livera

Jairus Brown
}
