% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\docType{data}
\name{qual_colours}
\alias{qual_colours}
\title{Qualitative Colour Palette}
\format{An object of class \code{character} of length 8.}
\usage{
qual_colours
}
\description{
Qualitative, colour blind safe colour palette developed by Okabe and Ito
}
\seealso{
\href{http://serialmentor.com/dataviz/color-pitfalls.html#not-designing-for-color-vision-deficiency}{serialmentor.com}
and \href{http://jfly.iam.u-tokyo.ac.jp/color/}{Okabe and Ito (2008)}
}
\author{
Benjamin R. Gordon
}
\keyword{datasets}
