% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\docType{data}
\name{sd_colours}
\alias{sd_colours}
\title{Sarah and Duck Colour Palette}
\format{An object of class \code{character} of length 5.}
\usage{
sd_colours
}
\description{
Colour palette derived from the TV show Sarah and Duck.
}
\seealso{
\href{https://www.sarahandduck.com}{sarahandduck.com}
}
\author{
Benjamin R. Gordon
}
\keyword{datasets}
