% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_palettes.R
\docType{data}
\name{warm_colours}
\alias{warm_colours}
\title{Warm Colour Palette}
\format{An object of class \code{character} of length 5.}
\usage{
warm_colours
}
\description{
Five sequential colours from the YlOrRd palette at colorbrewer.org.
}
\seealso{
\href{http://colorbrewer2.org/#type=sequential&scheme=YlOrRd&n=5}{colorbrewer.org}
}
\author{
Benjamin R. Gordon
}
\keyword{datasets}
